Added locking.
We use a very simple lock here -- just an integer and 1 and 0.
However, to work, we needed a yield syscall, which does nothing but give up the CPU, hopefully to someone
who has the lock. [Note: this would be a cool place to play with ticket transfer, etc]

We don't use any test-and-set, cas, etc atomic instructions in the lock functions; this is
safe because we are targetting a uniprocessor system and our scheduling in cooperative. The only
way we can get preempted is by making a syscall or instr/mem errors.

Notably, this makes malloc safe to use from more than one user processes at a time :)